Push

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/TelOp.java
@@ -271,7 +271,7 @@ case SAMPLEINTAKE:
 
 
  rbg.armrotatePIDF();
-            if(gamepad1.left_bumper) {
+            if(gamepad1.left_bumper && !rbg.flag[rbg.hang] && !rbg.flag[rbg.hang0]) {
                if (rbg.drop())  state = State.IDLE;
              else state = State.SAMPLEINTAKE;
             }
@@ -279,6 +279,8 @@ case SAMPLEINTAKE:
             if (rbg.timer(88000, rbg.start) || rbg.flag[rbg.force]) {
 
                 if (gamepad2.share || rbg.flag[rbg.hang]) {
+
+                    speed_factor = 1.0;
                     rbg.pre_hang();
                 }
                 if (gamepad1.share) {
